Initial secure checkers game baseline
This baseline has been setup to provide the following functionality:

1. Setup Spring MVC with controllers, models and views. Initial ability
allows users to view a mock profile + list other users that have
registered with the system.
2. Setup JPA to communicate with a database (a local H2 database saved
at ~/CheckersLocalPersistence.h2.db)
3. Setup an example performing Java Bean Validation in the User object
which limits the username to alphanumeric values ranging in length from
4-15 characters.
4. Setup Spring Security to require authentication to navigate the site
(see spring-security.xml)
5. Setup a StandardPasswordEncoder which salts+hashes user passwords to
save securely in the database
6. Load properties files with the possiblity of using jasypt encryption
7. Setup the Spring Test Framework to wire up the Spring context into
unit tests to make it easier to actually test various pieces of software

See the README for info on how to get the project running.