Initial crack at Game/Board object model
Not quite happy with this.. feel like plenty of good refactoring can be
done to clean this up. Just want to get this checked in to not hold up
any work @j13elliott may want to do. At this point you can create a new
Game, have a second player join the game, and the board is populated
with pieces.