Having the frontend actually manipulate the backend datastructure
Removed GameMovementControllerTestClass in favor of sending the data
into the GamesController. The backend logical board needed to be
inverted to map to how the frontend is configured. It is now updating
the backend with the appropriate data and saving the update game board.
(right now it is being saved twice per move, will fix that soon when I
have a little more time).